Add timeout to dock commands
It seems that dockutil is to fast for the ui to keep up with the
updates. Hence every playbook run removed/added only the first item
in the dock.
Through trial an error I came to the magic number of 7 seconds
between each dockutil remove or add commands. With such a timeout
it seems to work fine.